backend/requirements.txt
fastapi>=0.115.0,<1.0.0
pydantic>=2.7.0,<3.0.0
email-validator>=2.2.0,<3.0.0
uvicorn>=0.30.0,<1.0.0
sqlmodel>=0.0.22,<1.0.0
python-dotenv>=1.0.1,<2.0.0
authlib>=1.3.2,<2.0.0
openai>=1.52.0,<2.0.0
itsdangerous>=2.2.0,<3.0.0
pytest>=8.3.0,<9.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
